requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
pypdf==3.17.1
uvicorn[standard]==0.24.0
google-generativeai==0.5.0
pdf2image==1.16.3
Pillow==10.1.0
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
pytesseract==0.3.10
pymupdf==1.23.8
groq==0.11.0
